backend/app/requirements.txt
rasa==3.6.15
rasa-sdk==3.6.2
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
botocore==1.34.0
psycopg2-binary==2.9.9
pymongo==4.6.0

